Move metadata assets
