+++ tests/assets/sample_data/__init__.py
@@ -1,0 +1,5 @@
+# =====================================
+# Copyright: CEA-LIST/DIASI/SIALV/LVA
+# Author : pixano@cea.fr
+# License: CECILL-C
+# =====================================
